client-ocr/requirements.txt
pytesseract
pillow
numpy
opencv-python
pyautogui
requests